Update scrollbars.json (#167)
This update addresses a few places in roam where showing the scrollbar is also useful. More details here: paulovieira/roam-scrollbars#3

Meanwhile I found a related bug: in the settings panel, the main tab is not scrollable. See the screenshot here:

https://user-images.githubusercontent.com/2184309/188222688-68df7618-8a9a-48be-aa89-e20c24fbd1be.png

If the viewport is small I cannot reach the tabs for the extensions settings.