Update reference-path.json
Added a minimal commit to add a few more option in the offset settings. The rest is the same.

This is necessary to make this extension compatible with the Roam Studio extension, which changes the default css so much that the reference path is way off the center of the bullet. However it's just a mater of doing a correction with the "top offset" setting.