Update roam-preview-window.json
Updated the "json" file. 
Actually my extension of "mehboob elahi" folder got merged already, but since there is space in the folder and file name, that was causing an issued in installing the extension. So when I see the roam-depot directory to change the name of the folder and file, I notice that I cannot change the folder name i.e "mehboob elahi". 
I also found another account on my name "mehboobelahi05", so I updated this "json" file. Now creating PR for extenion in "mehboobelahi05" folder.
I request the admin to delete the "mehboob elahi" folder in the depot/extension, as that is a duplicated and causing error.
Thanks